# openspec/contracts/content-workflows/openapi/v1.yaml
openapi: 3.1.0
info:
  title: ACA Canonical Workflow API
  version: 2.0.0
  description: |
    Coordinated breaking replacement for ingestion, summarization, theme,
    digest, pipeline, podcast, and audio mutation contracts. All mutations
    return a durable OperationHandle backed by pgqueuer_jobs.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.aca.rotkohl.ai
    description: Production
security:
  - AdminKey: []
paths:
...
  /api/v1/uploads:
    post:
      operationId: createUpload
      summary: Upload a document for a subsequent files ingestion command
      tags: [uploads]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                title: { type: string }
                publication: { type: string }
            example:
              title: Architecture notes
              publication: Internal
      responses:
        "201":
          description: Durable upload reference
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadReference" }
              example:
                id: upl_01JABCDEF
                filename: architecture.pdf
                media_type: application/pdf
                size_bytes: 18432
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/ProblemResponse" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
  /api/v1/ingestions:
    get:
      operationId: listIngestionHistory
      summary: List compact terminal ingestion history derived from durable operations
      tags: [operations, ingestion]
      parameters:
        - name: command_key
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 100 }
        - name: configured_source_key
          in: query
          required: false
          schema: { type: string, pattern: "^src_[a-f0-9]{20}$" }
        - name: outcome
          in: query
          required: false
          schema: { $ref: "#/components/schemas/IngestionOutcome" }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/TerminalOperationStatus" }
        - name: parent_operation_id
          in: query
          required: false
          schema: { type: string, pattern: "^[1-9][0-9]*$", maxLength: 19 }
        - name: created_after
          in: query
          required: false
          schema: { type: string, format: date-time, maxLength: 64 }
        - name: created_before
          in: query
          required: false
          schema: { type: string, format: date-time, maxLength: 64 }
        - $ref: "#/components/parameters/PageLimit"
        - name: cursor
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 2048 }
          description: Signed keyset cursor bound to the normalized history filters.
      responses:
        "200":
          description: Compact ingestion history page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestionHistoryPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
    post:
      operationId: submitIngestion
      summary: Submit one typed ingestion command
      tags: [operations, ingestion]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IngestCommand" }
            example:
              kind: arxiv_paper
              identifier: "2401.12345"
              extract_pdf: true
              force_reprocess: false
      responses:
        "202":
          $ref: "#/components/responses/AcceptedIngestionOperation"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/ConflictProblem" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    PageLimit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 200 }
      example: "weekly-2026-07-06-gmail-rss"
...
  responses:
...
    AcceptedIngestionOperation:
      description: Durable ingestion accepted with the serving release identity
      headers:
        X-Release-Revision:
          required: true
          schema:
            type: string
            maxLength: 40
            pattern: "^(?:[a-f0-9]{40}|development|unavailable)$"
        X-Release-Revision-Source:
          required: true
          schema:
            type: string
            enum: [railway_commit_sha, local_development, unavailable]
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OperationHandle" }
    ProblemResponse:
      description: RFC 7807 problem
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/request-failed
            title: Request failed
            status: 400
            detail: The request could not be completed
    Unauthorized:
      description: Authentication is required
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Valid credentials are required
    Forbidden:
      description: Credentials do not authorize the operation
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/forbidden
            title: Forbidden
            status: 403
            detail: The supplied credentials are not authorized
    ValidationProblem:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/validation
            title: Validation failed
            status: 422
            detail: arxiv_paper.identifier is required
            errors:
              - { path: [identifier], code: missing, message: Field required }
    ConflictProblem:
      description: Current resource or operation state conflicts with the request
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/operation-conflict
            title: Operation conflict
            status: 409
            detail: Completed operations cannot be cancelled
...
  schemas:
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: [string, "null"] }
        code: { type: [string, "null"] }
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [path, code, message]
            properties:
              path: { type: array, items: { type: [string, integer] } }
              code: { type: string }
              message: { type: string }
    ResourceReference:
      type: object
      additionalProperties: false
      required: [type, id, url]
      properties:
        type:
          type: string
          enum: [content, ingestion_run, summary_batch, theme_analysis, digest, pipeline_run, podcast_script, podcast, audio_digest]
        id: { type: string }
        url: { type: string }
    IngestionOutcome:
      type: string
      enum: [success, zero_items, partial, failed, cancelled, unknown]
    OperationStatus:
      type: string
      enum: [queued, in_progress, completed, failed, cancelled]
    IngestionStatus:
      type: string
      enum: [ok, partial, error]
    TerminalOperationStatus:
      type: string
      enum: [completed, failed, cancelled]
...
    IngestionResult:
      oneOf:
        - { $ref: "#/components/schemas/IngestionResultV1" }
        - { $ref: "#/components/schemas/IngestionResultV2" }
      description: >-
        Legacy V1 rows may predate an explicit schema_version, so readers use
        structural union validation rather than a discriminator.
    IngestionResultV1:
      type: object
      additionalProperties: false
      required: [command_key, resolved_route, emitted_sources, items_ingested, content_ids]
      properties:
        schema_version: { type: integer, enum: [1], default: 1 }
        command_key: { type: string }
        resolved_route: { type: string }
        emitted_sources:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
        items_ingested: { type: integer, minimum: 0 }
        content_ids:
          type: array
          uniqueItems: true
          items: { type: integer, minimum: 1 }
        warnings: { type: array, items: { type: string } }
        details: { type: object, additionalProperties: true }
      description: >-
        Legacy reader shape. New writers never emit V1 and history must not
        infer missing outcome/count fields from it.
...
    PipelineResultV2:
      type: object
      additionalProperties: true
      x-operation-type: pipeline.run
      required: [schema_version, ingestion_summary]
      properties:
        schema_version: { type: integer, const: 2 }
        ingestion_summary: { $ref: "#/components/schemas/PipelineIngestionSummary" }
    ConfiguredSourceHistoryOutcome:
      type: object
      additionalProperties: false
      required: [source_key, status, outcome, items_ingested, items_failed]
      properties:
        source_key: { type: string, pattern: "^src_[a-f0-9]{20}$" }
        status: { $ref: "#/components/schemas/IngestionStatus" }
        outcome: { $ref: "#/components/schemas/IngestionOutcome" }
        items_ingested: { type: [integer, "null"], minimum: 0 }
        items_failed: { type: [integer, "null"], minimum: 0 }
        error_codes:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 100 }
        warning_codes:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 100 }
    IngestionHistoryItem:
      type: object
      additionalProperties: false
      required:
        - operation_id
        - command_key
        - operation_status
        - outcome
        - items_ingested
        - items_skipped
        - items_failed
        - source_outcomes
        - retry_count
        - status_url
        - created_at
      properties:
        operation_id: { type: string, pattern: "^[1-9][0-9]*$", maxLength: 19 }
        parent_operation_id: { type: [string, "null"], pattern: "^[1-9][0-9]*$", maxLength: 19 }
        command_key: { type: string, minLength: 1, maxLength: 100 }
        operation_status: { $ref: "#/components/schemas/TerminalOperationStatus" }
        outcome: { $ref: "#/components/schemas/IngestionOutcome" }
        items_ingested: { type: [integer, "null"], minimum: 0 }
        items_skipped: { type: [integer, "null"], minimum: 0 }
        items_failed: { type: [integer, "null"], minimum: 0 }
        source_outcomes:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/ConfiguredSourceHistoryOutcome" }
        retry_count: { type: integer, minimum: 0 }
        problem_code: { type: [string, "null"], maxLength: 100 }
        status_url: { type: string, pattern: "^/api/v1/operations/[1-9][0-9]*$" }
        created_at: { type: string, format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
    IngestionHistoryPage:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/IngestionHistoryItem" }
        next_cursor: { type: [string, "null"], maxLength: 2048 }
...
    OperationHandle:
      type: object
      additionalProperties: false
      x-operation-result-schemas:
        ingestion.execute: { $ref: "#/components/schemas/IngestionResult" }
        pipeline.run: { $ref: "#/components/schemas/PipelineResultV2" }
      required:
        - schema_version
        - operation_id
        - operation_type
        - status
        - progress
        - message
        - cancellable
        - retry_count
        - status_url
        - events_url
        - created_at
      properties:
        schema_version: { type: integer, const: 2 }
        operation_id: { type: string }
        operation_type:
          type: string
          enum:
            - ingestion.execute
            - summarization.run
            - theme_analysis.create
            - digest.create
            - pipeline.run
            - podcast_script.create
            - podcast_audio.create
            - audio_digest.create
        status: { $ref: "#/components/schemas/OperationStatus" }
        progress: { type: integer, minimum: 0, maximum: 100 }
        message: { type: string }
        cancellable: { type: boolean }
        retry_count: { type: integer, minimum: 0 }
        status_url: { type: string }
        events_url: { type: string }
        resource:
          anyOf:
            - { $ref: "#/components/schemas/ResourceReference" }
            - { type: "null" }
        result:
          type: [object, "null"]
          additionalProperties: true
        problem:
          anyOf:
            - { $ref: "#/components/schemas/Problem" }
            - { type: "null" }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
...
    UploadReference:
      type: object
      additionalProperties: false
      required: [id, filename, media_type, size_bytes]
      properties:
        id: { type: string }
        filename: { type: string }
        media_type: { type: string }
        size_bytes: { type: integer, minimum: 0 }
        title: { type: [string, "null"] }
        publication: { type: [string, "null"] }
...
    ConfiguredSourceCommandBase:
      type: object
      properties:
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
    GmailIngestCommand:
      allOf:
        - { $ref: "#/components/schemas/ConfiguredSourceCommandBase" }
        - type: object
          additionalProperties: false
          required: [kind]
          properties:
            kind: { type: string, const: gmail }
            configured_sources:
              type: array
              readOnly: true
              x-internal: true
              items: { type: object, additionalProperties: true }
            query: { type: string }
            max_items: { type: integer, minimum: 1 }
            days_back: { type: integer, minimum: 0 }
            after_date: { type: string, format: date-time }
            force_reprocess: { type: boolean, default: false }
    RssIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: rss }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
    BlogIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: blog }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
    SubstackIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: substack }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
    YouTubePlaylistIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: youtube_playlist }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
        public_only: { type: boolean, default: false }
    YouTubeRssIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: youtube_rss }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
    PodcastIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: podcast }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
        transcribe: { type: boolean, default: true }
    XSearchIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: x_search }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        prompt: { type: string }
        max_threads: { type: integer, minimum: 1 }
        force_reprocess: { type: boolean, default: false }
    PerplexitySearchIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: perplexity_search }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        prompt: { type: string }
        max_items: { type: integer, minimum: 1 }
        recency: { type: string, enum: [hour, day, week, month] }
        context_size: { type: string, enum: [low, medium, high] }
        force_reprocess: { type: boolean, default: false }
    FilesIngestCommand:
      type: object
      additionalProperties: false
      required: [kind, upload_ids]
      properties:
        kind: { type: string, const: files }
        upload_ids: { type: array, minItems: 1, items: { type: string } }
        force_reprocess: { type: boolean, default: false }
    UrlIngestCommand:
      type: object
      additionalProperties: false
      required: [kind, url]
      properties:
        kind: { type: string, const: url }
        url: { type: string, format: uri }
        title: { type: string }
        tags: { type: array, items: { type: string } }
        notes: { type: string }
        routing_mode: { type: string, enum: [auto, webpage], default: auto }
        force_reprocess: { type: boolean, default: false }
    ScholarSearchIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: scholar_search }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1, default: 20 }
    ScholarPaperIngestCommand:
      type: object
      additionalProperties: false
      required: [kind, identifier]
      properties:
        kind: { type: string, const: scholar_paper }
        identifier: { type: string, minLength: 1 }
        with_references: { type: boolean, default: false }
    ScholarReferencesIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: scholar_references }
        after: { type: string, format: date-time }
        before: { type: string, format: date-time }
        source_types: { type: array, items: { type: string } }
        dry_run: { type: boolean, default: false }
        limit: { type: integer, minimum: 1 }
    ArxivSearchIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: arxiv_search }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1, default: 20 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
        extract_pdf: { type: boolean, default: true }
    ArxivPaperIngestCommand:
      type: object
      additionalProperties: false
      required: [kind, identifier]
      properties:
        kind: { type: string, const: arxiv_paper }
        identifier: { type: string, minLength: 1 }
        extract_pdf: { type: boolean, default: true }
        force_reprocess: { type: boolean, default: false }
    HuggingFacePapersIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: huggingface_papers }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        max_items: { type: integer, minimum: 1, default: 30 }
        days_back: { type: integer, minimum: 0 }
        after_date: { type: string, format: date-time }
        force_reprocess: { type: boolean, default: false }
    ReadwiseIngestCommand:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, const: readwise }
        configured_sources:
          type: array
          readOnly: true
          x-internal: true
          items: { type: object, additionalProperties: true }
        updated_after: { type: string, format: date-time }
        source_types: { type: array, items: { type: string } }
        include_deleted: { type: boolean, default: false }
        max_books: { type: integer, minimum: 1 }
        force_reprocess: { type: boolean, default: false }
    IngestCommand:
      oneOf:
        - { $ref: "#/components/schemas/GmailIngestCommand" }
        - { $ref: "#/components/schemas/RssIngestCommand" }
        - { $ref: "#/components/schemas/BlogIngestCommand" }
        - { $ref: "#/components/schemas/SubstackIngestCommand" }
        - { $ref: "#/components/schemas/YouTubePlaylistIngestCommand" }
        - { $ref: "#/components/schemas/YouTubeRssIngestCommand" }
        - { $ref: "#/components/schemas/PodcastIngestCommand" }
        - { $ref: "#/components/schemas/XSearchIngestCommand" }
        - { $ref: "#/components/schemas/PerplexitySearchIngestCommand" }
        - { $ref: "#/components/schemas/FilesIngestCommand" }
        - { $ref: "#/components/schemas/UrlIngestCommand" }
        - { $ref: "#/components/schemas/ScholarSearchIngestCommand" }
        - { $ref: "#/components/schemas/ScholarPaperIngestCommand" }
        - { $ref: "#/components/schemas/ScholarReferencesIngestCommand" }
        - { $ref: "#/components/schemas/ArxivSearchIngestCommand" }
        - { $ref: "#/components/schemas/ArxivPaperIngestCommand" }
        - { $ref: "#/components/schemas/HuggingFacePapersIngestCommand" }
        - { $ref: "#/components/schemas/ReadwiseIngestCommand" }
      discriminator:
        propertyName: kind
        mapping:
          gmail: "#/components/schemas/GmailIngestCommand"
          rss: "#/components/schemas/RssIngestCommand"
          blog: "#/components/schemas/BlogIngestCommand"
          substack: "#/components/schemas/SubstackIngestCommand"
          youtube_playlist: "#/components/schemas/YouTubePlaylistIngestCommand"
          youtube_rss: "#/components/schemas/YouTubeRssIngestCommand"
          podcast: "#/components/schemas/PodcastIngestCommand"
          x_search: "#/components/schemas/XSearchIngestCommand"
          perplexity_search: "#/components/schemas/PerplexitySearchIngestCommand"
          files: "#/components/schemas/FilesIngestCommand"
          url: "#/components/schemas/UrlIngestCommand"
          scholar_search: "#/components/schemas/ScholarSearchIngestCommand"
          scholar_paper: "#/components/schemas/ScholarPaperIngestCommand"
          scholar_references: "#/components/schemas/ScholarReferencesIngestCommand"
          arxiv_search: "#/components/schemas/ArxivSearchIngestCommand"
          arxiv_paper: "#/components/schemas/ArxivPaperIngestCommand"
          huggingface_papers: "#/components/schemas/HuggingFacePapersIngestCommand"
          readwise: "#/components/schemas/ReadwiseIngestCommand"
...